pyproject.toml
[project]
name = "cerbottana"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aiohttp==3.13.2",
    "domify==0.4.7",
    "imageprobe==0.3.0",
    "pillow==12.0.0",
    "pokedex[asyncio] @ https://github.com/Parnassius/pokedex.git",
    "python-dateutil==2.9.0.post0",
    "sqlalchemy==2.0.44",
    "typenv==0.2.0",
]

[project.scripts]
cerbottana = "cerbottana.__main__:main"

[dependency-groups]
dev = [
    "darglint==1.8.1",
    "freezegun==1.5.5",
    "mypy==1.18.2",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-cov==7.0.0",
    "pytest-mock==3.15.1",
    "python-dotenv[cli]==1.2.1",
    "ruff==0.14.3",
    "types-python-dateutil==2.9.0.20251008",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:"]

[tool.coverage.run]
relative_files = true
source = ["src/cerbottana"]


[tool.hatch.metadata]
allow-direct-references = true


[tool.mypy]
python_version = "3.14"
strict = true
implicit_reexport = true

# Disallow dynamic typing
disallow_any_unimported = true
#disallow_any_expr = true
#disallow_any_decorated = true
disallow_any_explicit = true

show_error_codes = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_any_explicit = false
disallow_untyped_defs = false
disallow_incomplete_defs = false


[tool.pytest.ini_options]
addopts = "--strict-config --strict-markers"
markers = ["integration"]
asyncio_mode = "auto"
xfail_strict = true


[tool.ruff.lint]
select = [
    "A",     # flake8-builtins
    "ANN",   # flake8-annotations
    "ARG",   # flake8-unused-arguments
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "E",     # pycodestyle
    "EM",    # flake8-errmsg
    "F",     # pyflakes
    "FURB",  # refurb
    "I",     # isort
    "ICN",   # flake8-import-conventions
    "INP",   # flake8-no-pep420
    "ISC",   # flake8-implicit-str-concat
    "N",     # pep8-naming
    "PERF",  # perflint
    "PGH",   # pygrep-hooks
    "PT",    # flake8-pytest-style
    "PTH",   # flake8-use-pathlib
    "RET",   # flake8-return
    "RSE",   # flake8-raise
    "RUF",   # Ruff-specific rules
    "SIM",   # flake8-simplify
    "SLF",   # flake8-self
    "T10",   # flake8-debugger
    "TC",    # flake8-type-checking
    "TID",   # flake8-tidy-imports
    "TRY",   # tryceratops
    "UP",    # pyupgrade
    "W",     # pycodestyle
    "YTT",   # flake8-2020
]
ignore = [
    "A003",  # builtin-attribute-shadowing
    "TC001", # typing-only-first-party-import
    "TC002", # typing-only-third-party-import
    "TC003", # typing-only-standard-library-import
]
fixable = [
    "I",     # isort
    "F401",  # unused-import
    "TC",    # flake8-type-checking
    "TID",   # flake8-tidy-imports
    "UP",    # pyupgrade
]
dummy-variable-rgx = "^_$"

[tool.ruff.lint.per-file-ignores]
"src/cerbottana/plugins/*" = ["A005"]  # stdlib-module-shadowing
"tests/*" = [
    "ANN",  # flake8-annotations
    "ARG001",  # unused-function-argument
    "SLF",  # flake8-self
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "cerbottana.typedefs.RoomId",
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
split-on-trailing-comma = false
